dorkenum_package_v2_v5/requirements.txt
click
requests
PyYAML
jinja2
pytest
